requirements.txt
discord.py==2.0.0
slash_util>=1.2.0
aiosqlite>=0.17
async_google_trans_new>=1.4.3
motor>=2.5.1
requests>=2.27.0
